Prepends the label of the select component as part of the aria-label.
Fixes CAB-3803.

According to WC3, the aria-label attribute is meant to read the label of the listbox. With commit 35bae7ff1459147ad94affa028119517fc1b6b71,
the aria-label was changed to read the selected option's value, which is not what this attribute is meant to be used for. This commit changes
the aria-label to first display the standard label name, and then appends the selected option value. Even though it is not standard to use this
attribute to read the selected value, it is meant to mitigate the problem with mat-select where the selected option is not read at all.